refactor: application.mangle -> application.url 로 패키지 이름 변경
