+++ src/main/java/io/github/daengdaenglee/mangurl/application/url/service/Base62EncodeService.java
@@ -1,4 +1,4 @@
-package io.github.daengdaenglee.mangurl.application.mangle.service;
+package io.github.daengdaenglee.mangurl.application.url.service;
 
 import org.springframework.stereotype.Service;
