+++ src/test/java/io/github/daengdaenglee/mangurl/outboundadapter/mangle/repository/InMemoryUrlRepositoryTest.java
@@ -1,6 +1,6 @@
 package io.github.daengdaenglee.mangurl.outboundadapter.mangle.repository;
 
-import io.github.daengdaenglee.mangurl.application.mangle.outboundport.DuplicateShortUrlCodeException;
+import io.github.daengdaenglee.mangurl.application.url.outboundport.DuplicateShortUrlCodeException;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
